# docs/swagger.yaml
basePath: /v1
definitions:
  api.HTTPError:
    properties:
      message: {}
    type: object
  model.Queue:
    properties:
      current_number:
        minimum: 0
        type: integer
      domain:
        type: string
      permitted_number:
        minimum: 0
        type: integer
    required:
    - domain
    type: object
  model.WhiteList:
    properties:
      domain:
        type: string
    required:
    - domain
    type: object
info:
  contact: {}
  description: API for WaitingRoom
  license:
    name: MIT
    url: http://github.com/pyama86/ngx_waitingroom/LICENSE
  title: WaitingRoomAPI
  version: "1.0"
paths:
  /queues:
    get:
      consumes:
      - application/json
      description: get queues
      operationId: queues#get
      parameters:
      - description: Queue Domain
        in: query
        name: domain
        type: string
      - description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: per_page
        in: query
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Queue'
            type: array
        "404":
          description: Not Found
          schema:
            items:
              $ref: '#/definitions/model.Queue'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: get queues
      tags:
      - queues
    post:
      consumes:
      - application/json
      description: create queue
      operationId: queues#post
      parameters:
      - description: Queue Object
        in: body
        name: queue
        required: true
        schema:
          $ref: '#/definitions/model.Queue'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: create queue
      tags:
      - queues
  /queues/{domain}:
    delete:
      consumes:
      - application/json
      description: delete queue
      operationId: queues#delete
      parameters:
      - description: Queue Name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: delete queue
      tags:
      - queues
    put:
      consumes:
      - application/json
      description: update queue
      operationId: queues#put
      parameters:
      - description: Queue Name
        in: path
        name: domain
        required: true
        type: string
      - description: Queue Object
        in: body
        name: queue
        required: true
        schema:
          $ref: '#/definitions/model.Queue'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: update queue
      tags:
      - queues
  /viron:
    get:
      consumes:
      - application/json
      description: get global menu
      operationId: viron#get
      produces:
      - application/json
      responses: {}
      summary: get global menu
      tags:
      - viron
  /viron_authtype:
    get:
      consumes:
      - application/json
      description: get auth type
      operationId: viron_authtype#get
      produces:
      - application/json
      responses: {}
      summary: get auth type
      tags:
      - viron
  /whitelist:
    get:
      consumes:
      - application/json
      description: get whiteLists
      operationId: whitelist#get
      parameters:
      - description: WhiteList Domain
        in: query
        name: domain
        type: string
      - description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: per_page
        in: query
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.WhiteList'
            type: array
        "404":
          description: Not Found
          schema:
            items:
              $ref: '#/definitions/model.WhiteList'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: get whiteLists
      tags:
      - whitelist
    post:
      consumes:
      - application/json
      description: create whiteList
      operationId: whitelist#post
      parameters:
      - description: WhiteList Object
        in: body
        name: whitelist
        required: true
        schema:
          $ref: '#/definitions/model.WhiteList'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: create whiteList
      tags:
      - whitelist
  /whitelist/{domain}:
    delete:
      consumes:
      - application/json
      description: delete whiteList
      operationId: whitelist#delete
      parameters:
      - description: WhiteList Domain
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: delete whiteList
      tags:
      - whitelist
swagger: "2.0"
tags:
- name: queues
- name: whitelist
- name: viron